requirements.txt
# Pitch Generator - Dependencies

# Web Framework
streamlit>=1.28.0

# HTTP Requests
requests>=2.31.0

# Google Gemini API
google-generativeai>=0.3.0

# Image Processing
Pillow>=10.0.0

# PDF Generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
